format url api module (#34111)
* format url api module

* Bump pack from version CommonScripts to 1.14.39.

---------
