+++ Packs/CommonScripts/ReleaseNotes/1_14_39.md
@@ -1,0 +1,7 @@
+
+#### Scripts
+
+##### FormatURL
+- Updated the Docker image to: *demisto/python3:3.10.14.92207*.
+- Internal code adjustments.
+
