+++ Packs/CommonScripts/Scripts/FormatURL/FormatURL.yml
@@ -18,7 +18,7 @@ tags:
 timeout: '0'
 type: python
 subtype: python3
-dockerimage: demisto/python3:3.10.13.89009
+dockerimage: demisto/python3:3.10.14.92207
 fromversion: 5.5.0
 tests:
 - FormatURL-Test
