+++ Packs/Phishing/.secrets-ignore
@@ -7,3 +7,4 @@ u003cDB6PR0701MB29994477F1C63A272139C87ECF679@DB6PR0701MB2999.eurprd07.prod.outl
 DB6PR0701MB2999.eurprd07.prod.outlook.com
 2603:10a6:4:74::13
 2603:10a6:10:181::15
+http://hellothere.com
