+++ Packs/Phishing/ReleaseNotes/3_6_19.md
@@ -1,0 +1,7 @@
+
+#### Scripts
+
+##### FindDuplicateEmailIncidents
+- Updated the Docker image to: *demisto/sklearn:1.0.0.92988*.
+- Fixed an issue where the script timed out due to too many calls to the ***FormatURL*** script.
+
